requirements.txt
streamlit>=1.0
pandas>=1.0
numpy
faker>=15.0
openpyxl
xlsxwriter
google.genai>=0.1.0
matplotlib.pyplot